backend/requirements.txt
# SmartCity Hackathon Backend - core (Python 3.8+)
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
python-dotenv==1.0.1
numpy>=1.20,<1.25
aiosqlite==0.19.0
httpx==0.27.0

# Optional (Python 3.9+): uncomment for Gemini Vision + embeddings
# google-generativeai>=0.4.0
# opencv-python-headless>=4.8.0
